+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+# [1.2.0](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/compare/v1.1.0...v1.2.0) (2025-07-10)
+
+
+### Features
+
+* add example for new credit card tokenizer sdk ([a774ed6](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/a774ed6c54768748fe13f4925de41393097f7e07))
+* update PCPCreditCardTokenizer to use new Hosted Tokenization SDK ([47373e9](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/47373e98c3661342c81ac15942da306ceddfea5d))
+
+
+### Reverts
+
+* Revert "chore: update SonarCloud action" ([907c3f9](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/907c3f9a4b4b19232d173d6e5f2b942d4cda0a69))
+
+
+
+# [1.1.0](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/compare/v1.0.0...v1.1.0) (2025-07-03)
+
+
+### Features
+
+* add Google Pay button interface ([2bea28e](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/2bea28ee52b954a726d58b0528bc163d2742728a))
+* add Google Pay demo implementation for web components ([f67ae0d](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/f67ae0d8662c2443a640b9a11112448697b7e085))
+* add listener for payment data changes to example ([3a8acf1](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/3a8acf10366f8c63e6c77480af2af152aca158ed))
+* enhance example with additional event handlers and types ([d0b9abb](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/commit/d0b9abb71154fc0cc046128919c38c868cfe26e2))
+
+
+
 # [1.0.0](https://github.com/PAYONE-GmbH/PCP-client-javascript-SDK/compare/v0.0.2...v1.0.0) (2024-08-22)